+++ dart_sample/lib/sample_routes.dart
@@ -145,6 +145,7 @@ class SampleRoutes {
   static const String overflowIndicator = '/overflow-indicator';
   static const String offstage = '/offstage';
   static const String baseline = '/baseline';
+  static const String safeArea = '/safe-area';
   static const String richText = '/rich-text';
   static const String intrinsicWidgets = '/intrinsic-widgets';
   static const String layoutBuilder = '/layout-builder';